separate tests into unit and base test files
